chore(workspace): move repository adapter
- keep workspace persistence implementation inside the workspace module

- leave shared MongoDB schema ownership unchanged
